Refactor key management and add master key functionality Update dependencies and implement Tauri integration Update dependencies and data access layer for OAuth user consent Refactor authentication state and add UserInfo endpoint support

+++ crates/model/src/contract/jwk_private.rs
@@ -24,7 +24,7 @@ pub struct JwkPrivate {
     #[serde(rename = "use")]
     pub r#use: KeyUse,
 
-    pub kid: i64,
+    pub kid: u32,
 
     /// Algorithm intended for use
     pub alg: JwsAlgorithm,
