+++ crates/model/src/contract/jwk_public.rs
@@ -23,7 +23,7 @@ pub struct JwkPublic {
     pub r#use: KeyUse,
 
     /// Key ID
-    pub kid: i64,
+    pub kid: u32,
 
     /// Algorithm intended for use
     pub alg: JwsAlgorithm,
