+++ crates/service/src/repo/oauth2_authorization_code_repo.rs
@@ -7,7 +7,7 @@ pub trait OAuth2AuthorizationCodeRepo {
     fn create_authorization_code(
         &self,
         client_id: String,
-        key_id: i64,
+        key_id: u32,
         redirect_uri: String,
         scopes: Vec<String>,
         resource: Option<String>,
